Initial commit
This is a mash up of scripting that I put together, with help from my DBA, back at a startup to accurately gauge MSSQL server performance ahead of migrations. There are assumptions made in the base of the code on IO size and data access patterns; these can be easily changed to meet your needs.